+++ CHANGELOG.md
@@ -9,23 +9,36 @@ All notable changes to this project will be documented in this file.
 ```ts
 const logic = kea([
   actions({
-    setUser: (user) => ({ user }),
     fetchUser: (id: number) => ({ id }),
   }),
-  reducers({ user: { setUser: (_, { user }) => user } }),
   listeners({
     fetchUser: async ({ id }, breakpoint) => {
+      await breakpoint(100)
       const user = await fetch(`https://example.com/users/${id}`)
       breakpoint()
-      actions.setUser(user)
+      return user
     },
   }),
 ])
-await logic.actions.fetchUser(1)
-console.log(logic.values.user)
+const user = await logic.actions.fetchUser(1)
 ```
 
-To make this work, each created action now also comes with a `dispatchId`:
+The promise returns whatever is returned in the _first listener_ that listens to this action. Ususally
+that's the output of the only listener is the same logic that creates the action.
+
+In case you use breakpoints, and the action is called multiple times, all the promises will resolve 
+when the last called action returns.
+
+That means in the case of
+
+```ts
+const promise1 = logic.actions.fetchUser(1)
+const promise2 = logic.actions.fetchUser(1)
+```
+
+Both promises will resolve at the same time. The first dispatch one that breaks will resolve when the second one finishes.
+
+To make this work, each created action now also comes with an ever-increasing `dispatchId`:
 
 ```js
 logic.actionCreators.fetchUser(123) === {
@@ -35,7 +48,8 @@ logic.actionCreators.fetchUser(123) === {
 }
 ```
 
-This can be disabled by calling `resetContext({ disableDispatchId: true })`.
+To disable setting `dispatchId` and hence support for async actions, call `resetContext({ disableAsyncActions: true })`.
+
 
 ## 3.1.0 - 2022-12-13